requirements.txt
robotframework>=6.1.1
robotframework-seleniumlibrary>=6.1.3
selenium>=4.15.2
webdriver-manager>=4.0.1
robotframework-pabot>=2.16.0
robotframework-requests>=0.9.5
robotframework-pythonlibcore>=4.2.0
pytest>=7.4.3
python-dotenv>=1.0.0
allure-robotframework>=2.14.2 